+++ frontend/src/pages/dev/editor/fileHandler/csvHandler.ts
@@ -4,20 +4,26 @@ import { Polygon } from '@/mod/data/polygon/polygon';
 import type { Coord } from '@/mod/data/com/vertex';
 
 export function parseCSV(fileData: string) {
-    const nestedArrayData: string[][] = fileData.split('\n').map((line) => line.split(','));
+	console.log(fileData)
+    const nestedArrayData: string[][] = fileData.split('\n').map((line) => line.split(';'));
     // Remove the last line due to trailing \n
     nestedArrayData.pop();
+	console.log(nestedArrayData)
 
     const roomDataArray: Room[] = [];
 
     for (const dataLine of nestedArrayData) {
         const id: string = dataLine[0];
         const vertices: Coord[] = [];
-        for (let idx = 1; idx < dataLine.length; idx += 2) {
-            vertices.push([parseFloat(dataLine[idx]), parseFloat(dataLine[idx + 1])]);
+        for (let idx = 1; idx < dataLine.length; idx++) {
+			const splitCoord = dataLine[idx].split(',')
+            vertices.push([parseFloat(splitCoord[0]), parseFloat(splitCoord[1])]);
         }
+		console.log(id)
+		console.log(vertices)
         const roomPolygon: Polygon = new Polygon(vertices, [70, 70, 70, 0.4]);
-        const roomData: Room = new Room(0, 0, roomPolygon, id);
+        const roomData: Room = new Room(0, 1, roomPolygon, id);
+		console.log(roomData)
         roomDataArray.push(roomData);
     }