+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-# Pathfinding
-- multi floor pathfinding done
+# Mappy
+- mappy cli now can add vertex to polygon
+
+# Frontend
+- changed csv parser to handle the new csv format